+++ src/main/java/hudson/plugins/libvirt/lib/ConnectionBuilder.java
@@ -1,11 +1,13 @@
 package hudson.plugins.libvirt.lib;
 
-import com.cloudbees.plugins.credentials.common.StandardCredentials;
+import hudson.plugins.libvirt.lib.libvirt.LibvirtConnectAuth;
+import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
 import hudson.plugins.libvirt.lib.libvirt.LibVirtConnectImpl;
 import java.util.Locale;
 import java.util.logging.Level;
 import java.util.logging.LogRecord;
 import java.util.logging.Logger;
+import org.libvirt.ConnectAuth;
 
 /**
  * Created by magnayn on 05/02/2014.
@@ -17,10 +19,13 @@ public class ConnectionBuilder {
     private boolean readOnly = false;
 
     private String hypervisorType;
+    private String hypervisorTransport;
     private String userName;
     private String hypervisorHost;
     private int    hypervisorPort;
     private String hypervisorSysUrl;
+
+    private StandardUsernamePasswordCredentials credentials;
 
     public static ConnectionBuilder newBuilder() {
         return new ConnectionBuilder();
@@ -28,6 +33,11 @@ public class ConnectionBuilder {
 
     public ConnectionBuilder hypervisorType(String type) {
         this.hypervisorType = type;
+        return this;
+    }
+
+    public ConnectionBuilder hypervisorTransport(String transport) {
+        this.hypervisorTransport = transport;
         return this;
     }
 
@@ -51,8 +61,8 @@ public class ConnectionBuilder {
         return this;
     }
 
-    public ConnectionBuilder withCredentials(StandardCredentials standardCredentials) {
-        // TODO implement credentials
+    public ConnectionBuilder withCredentials(StandardUsernamePasswordCredentials userPwCredentials) {
+        this.credentials = userPwCredentials;
         return this;
     }
 
@@ -66,19 +76,32 @@ public class ConnectionBuilder {
         return this;
     }
 
+    public boolean isLibraryTransport() {
+        return hypervisorTransport != null && hypervisorTransport.startsWith("libssh");
+    }
+
     public IConnect build() throws VirtException {
         if (uri == null) {
             uri = constructHypervisorURI();
         }
-
-        return new LibVirtConnectImpl(uri, readOnly);
+        ConnectAuth auth = null;
+        if (isLibraryTransport() && credentials != null) {
+            auth = new LibvirtConnectAuth(credentials, userName);
+        }
+        return new LibVirtConnectImpl(uri, auth, readOnly);
     }
 
     public String constructHypervisorURI() {
-        String url = hypervisorType.toLowerCase(Locale.ENGLISH) + "://";
+        String url = hypervisorType.toLowerCase(Locale.ENGLISH);
+        if (hypervisorTransport != null && !hypervisorTransport.isEmpty()) {
+            url += "+" + hypervisorTransport;
+        }
+        url += "://";
         // Fixing JENKINS-14617
         if (userName != null && !userName.isEmpty()) {
             url += userName + "@";
+        } else if (credentials != null) {
+            url += credentials.getUsername() + "@";
         }
 
         url += hypervisorHost;
@@ -88,6 +111,23 @@ public class ConnectionBuilder {
 
         if (hypervisorSysUrl != null && !hypervisorSysUrl.isEmpty()) {
             url += "/" + hypervisorSysUrl;
+            if (!hypervisorSysUrl.contains("?")) {
+                url += "?";
+            }
+        } else {
+            url += "?";
+        }
+        if ((hypervisorSysUrl == null || !hypervisorSysUrl.contains("no_tty="))
+                && "ssh".equals(hypervisorTransport)) {
+            url += "no_tty=1";
+        }
+        if ((hypervisorSysUrl == null || !hypervisorSysUrl.contains("sshauth="))
+                && isLibraryTransport() && credentials != null) {
+            url += "&sshauth=password";
+        }
+        if ((hypervisorSysUrl == null || !hypervisorSysUrl.contains("known_hosts_verify="))
+                && isLibraryTransport()) {
+            url += "&known_hosts_verify=auto";
         }
 
         LogRecord rec = new LogRecord(Level.INFO, "hypervisor: {0}");